requirements.txt
fastapi==0.115.14
fastapi==0.115.14
passlib==1.7.4
pydantic==2.11.7
python_jose==3.5.0
SQLAlchemy==2.0.41
